requirements.txt
absl-py==0.7.1
astor==0.8.0
dotmap==1.3.8
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
mock==3.0.5
numpy==1.16.3
pandas==0.24.2
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tqdm==4.32.1
Werkzeug==0.15.4
